Move testing stuff into test scope
